Lift `FontResource` common to `appleMain` with iOS/macOS extensions.
The `FontResource` code is the same on both iOS and macOS, except for the helper to create either a `UIFont` or an `NSFont`. Move those helpers to platform-specific code, while keeping the common code shared for easier maintenance.

This follows the same approach that we use for `ColorResource`.
